Reorganized the project (mostly specs)
I've always found it best to keep the spec file structure the same as
the library file structure. It makes it easy to see what has/hasn't
been spec'd, where specific specs should live, etc. This moves
some existing specs into the appropriate place and creates
placeholders for others.

This also moves the imap extensions out into the Gmail namespace
which makes a bit more sense then just having it be a standalone
class living in the /gmail/client folder.

Sidenote: We could probably better split out the specs and DRY them
up a bit. Especially the individual client specs.
