+++ lib/gmail/client.rb
@@ -25,7 +25,7 @@ module Gmail
       end
     end
 
-    require 'gmail/client/imap_extensions'
+    require 'gmail/imap_extensions'
     require 'gmail/client/base'
     require 'gmail/client/plain'
     require 'gmail/client/xoauth'
