+++ lib/gmail/client/base.rb
@@ -24,7 +24,7 @@ module Gmail
       # Connect to gmail service.
       def connect(raise_errors = false)
         @imap = Net::IMAP.new(GMAIL_IMAP_HOST, GMAIL_IMAP_PORT, true, nil, false)
-        GmailImapExtensions.patch_net_imap_response_parser
+        Gmail::ImapExtensions.patch_net_imap_response_parser
         @imap
       rescue SocketError
         raise_errors and raise ConnectionError, "Couldn't establish connection with Gmail IMAP service"
