+++ spec/gmail/client/base_spec.rb
@@ -1,0 +1,5 @@
+require 'spec_helper'
+
+describe Gmail::Client::Base do
+  skip "add some examples to (or delete) #{__FILE__}"
+end
