+++ spec/gmail/client/plain_spec.rb
