+++ spec/gmail/client/xoauth_spec.rb
@@ -1,0 +1,5 @@
+require 'spec_helper'
+
+describe Gmail::Client::XOAuth do
+  skip "add some examples to (or delete) #{__FILE__}"
+end
