+++ spec/gmail/imap_extensions_spec.rb
@@ -1,11 +1,11 @@
 require 'rubygems'
 require 'rspec'
 require 'net/imap'
-require 'gmail/client/imap_extensions'
+require 'gmail/imap_extensions'
 
-describe 'imap patch' do
+describe Gmail::ImapExtensions do
   it 'does not modify arity' do
-    GmailImapExtensions::patch_net_imap_response_parser
+    Gmail::ImapExtensions::patch_net_imap_response_parser
     expect(Net::IMAP::ResponseParser.new).to respond_to(:msg_att).with(0).arguments
     expect(Net::IMAP::ResponseParser.new).to respond_to(:msg_att).with(1).arguments
   end
